simplify ssr bundle cuts: module boundaries instead of inlined loaders
Review pass over the SSR-shrink changes, keeping the same server-bundle
wins but with less machinery:

- orpc client: back to the in-process router client on the server. The
  worker serves /api itself (routes/api.orpc.$.ts statically imports
  appRouter), so the router and every domain handler are in the worker
  bundle no matter what client.ts imports. The OpenAPI loopback bought
  zero bytes while adding a self-HTTP hop per SSR call plus cookie and
  base-URL forwarding that had already needed two fixup commits.

- source-code-block / serialized-object-code-block: the implementations
  move verbatim to *.client.tsx with their original static imports and
  full CodeMirror typing; the public modules become the same thin
  SSR-aware lazy wrapper already used by project-stream-view.lazy.tsx.
  This drops the any/unknown type erasure, the per-prop extension
  threading, and the editorReadySignal workaround: with a module
  boundary the editor mounts in the same commit as its parent again, so
  the existing rAF scroll-to-bottom just works.

- posthog: nothing in os/semaphore consumes the PostHog React context,
  so the provider (and the async provider swap, which remounted the
  entire app tree when the fragment was replaced by the loaded provider)
  is gone. PostHogInit is a null-rendering component that lazy-loads
  posthog-js in the browser and calls init; posthog-js/react is no
  longer imported at all.
